BQProxy/requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cloudevents==1.12.0
deprecation==2.1.0
Flask==3.1.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud==0.34.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
requests==2.32.4
rsa==4.9.1
six==1.17.0
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
